MIDOGpp-main/requirements.txt
numpy<1.24.0,>=1.21.0
torch==1.13.0
torchvision==0.14.0
torchaudio==0.13.0
fastai==1.0.61
pandas<2.0.0,>=1.3.0
evalutils==0.3.1
hydra-core==1.3.2
object_detection_fastai==0.0.10
omegaconf==2.2.3
opencv-python-headless==4.5.5.64
openslide-python==1.1.2
Pillow==9.4.0
PyYAML==6.0
scikit-learn==1.0.2
SimpleITK==2.2.1
SlideRunner==2.0
torchmetrics==0.11.0
tqdm>=4.60.0
albumentations==1.3.1
tiatoolbox==1.0.1
wandb==0.12.4
protobuf==3.20.3